Move methods not related to Vert.x out of VertxUtil class
